refactor(breadcrumb separator): Refactored breadcrumb separator rendering in theme switcher example
# Project State

## Current Focus
Refactored breadcrumb separator rendering in theme switcher example

## Context
The change simplifies breadcrumb separator rendering by removing redundant type conversions in the index calculation

## Completed
- [x] Removed redundant `as usize` conversion in breadcrumb separator index calculation
- [x] Updated Cargo.lock with dependency version bump

## In Progress
- [x] No active work in progress

## Blockers
- None

## Next Steps
1. Review other similar index calculations in the theme switcher example
2. Consider additional refactoring opportunities in the breadcrumb rendering logic
---
category: refactor
scope: breadcrumb separator
---

+++ examples/theme_switcher.rs
@@ -800,7 +800,7 @@ impl Widget for WidgetDemoPanel {
         let mut x = start_x;
         for (i, (crumb, is_last)) in crumbs.iter().enumerate() {
             if i > 0 {
-                let idx = breadcrumb_row as usize * area.width as usize + x as usize;
+                let idx = breadcrumb_row * area.width as usize + x as usize;
                 if idx < plane.cells.len() {
                     plane.cells[idx].char = '›';
                     plane.cells[idx].fg = theme.fg_muted;
@@ -809,7 +809,7 @@ impl Widget for WidgetDemoPanel {
             }
 
             for c in crumb.chars() {
-                let idx = breadcrumb_row as usize * area.width as usize + x as usize;
+                let idx = breadcrumb_row * area.width as usize + x as usize;
                 if idx < plane.cells.len() {
                     plane.cells[idx].char = c;
                     plane.cells[idx].fg = if *is_last { theme.primary } else { theme.fg };
